Remove deprecated SubprocessToString files and update usage
