+++ src/base/SubprocessUtils.hpp
@@ -5,6 +5,22 @@
 
 namespace et {
 /**
+ * @brief Executes a shell command and captures all stdout into a string.
+ * @param cmd Null-terminated command string passed to popen.
+ */
+inline std::string SystemToStr(const char* cmd) {
+  std::array<char, 128> buffer;
+  std::string result;
+  std::shared_ptr<FILE> pipe(popen(cmd, "r"), pclose);
+  if (!pipe) throw std::runtime_error("popen() failed!");
+  while (!feof(pipe.get())) {
+    if (fgets(buffer.data(), 128, pipe.get()) != nullptr)
+      result += buffer.data();
+  }
+  return result;
+}
+
+/**
  * @brief Utility class for executing subprocesses and capturing output.
  */
 class SubprocessUtils {
